Makes folder for Containers (stateful) and Components (stateless)
Moves route transition handler to file for route-handling Container
