+++ source/components/FourOhFour/index.js
@@ -1,0 +1,12 @@
+import React from 'react'
+
+import Main from '../../layouts/Main'
+import styles from './styles.css'
+
+export default () => (
+  <Main title="404">
+    <section className={ styles.FourOhFour }>
+      <div>Whoops!</div>
+    </section>
+  </Main>
+)
